The first iteration of my Etch-a-sketch website
By default, a 16 x 16 grid (canvas) dynamically added to the page.
The grid is wired with a mouseover event listener that handles the event
by changing the color of the grid element that dispatached the event.

The 'New Grid' button renders a new grid of the size specified by the
user.