Modified grid appearance and added a square limit
This iteration of etch-sketch limits the grid size to 100 x 100. If the
user enters a value greater than 100, the grid will default to the
aforementioned limit.
To improve the grid appearance, I applied padding to the grid container,
added a thin border to each square in the grid and reduced the thicknes
of the grid border.